Notification implementation / cleanup files
